IDs type is int64

+++ database_base_model.go
@@ -5,7 +5,7 @@ import "time"
 // gorm.Model alternative without DeletedAt column (to disable Soft Delete)
 // see https://gorm.io/docs/delete.html#Soft-Delete
 type BaseModel struct {
-	ID        uint `gorm:"primarykey"`
+	ID        int64 `gorm:"primarykey"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
